+++ CHANGES.rst
@@ -9,6 +9,8 @@ Changelog
 -   Allow iDQ to label DQV onto events based on p(glitch). 
     Adjustable by pipeline.
 
+-   Move functions in tasks/lalinference.py to lalinference_pipe.py in lalsuite
+
 0.3.1 (2019-03-18)
 ------------------